change in a plugins
Добавил плагин image - все методы для работы с картинками нужно
реализовывать в нем

плагин dir переименован в setting - в этом модуле вся работа по
настройке справочника.

docs переименован в doc

On branch snakejet
Your branch is up-to-date with 'origin/snakejet'.
Changes to be committed:
	new file:   rpc_plugin/doc.js
	deleted:    rpc_plugin/docs.js
	new file:   rpc_plugin/image.js
	renamed:    rpc_plugin/dir.js -> rpc_plugin/setting.js
	modified:   rpc_plugin/user.js
